Steps to convert account imap functions to c++
These changes separate out the account imap functions to a new file with
the intention to change to c++
